+++ packages/telescope-releases/lib/server/import_releases.js
@@ -13,6 +13,14 @@ importRelease = function (number) {
 };
 
 Meteor.startup(function () {
+  
   importRelease('0.11.0');
   importRelease('0.11.1');
+  importRelease('0.11.2');
+
+  // if this is before the first run, mark all release notes as read to avoid showing them
+  if (!Events.findOne({name: 'firstRun'})) {
+    var r = Releases.update({}, {$set: {read: true}}, {multi: true});
+  }
+
 });