+++ packages/telescope-blank/lib/custom_fields.js
@@ -8,7 +8,8 @@ var customProperty = {
     optional: true,                         // make this property optional
     autoform: {
       editable: true,                       // make this property editable by users
-      type: "bootstrap-datetimepicker"      // assign a custom input type
+      type: "bootstrap-datetimepicker",     // assign a custom input type
+      omit: false                           // set to true to omit field from form entirely
     }
   }
 }