+++ packages/telescope-releases/lib/server/import_releases.js
@@ -23,7 +23,7 @@ Meteor.startup(function () {
   
   importRelease('0.11.0');
   importRelease('0.11.1');
-  importRelease('0.11.2');
+  importRelease('0.12.0');
 
   // if this is before the first run, mark all release notes as read to avoid showing them
   if (!Events.findOne({name: 'firstRun'})) {